bot/requirements.txt
discord.py
python-dotenv
httpx
pydantic
mcrcon
